refactor(preflight): unit_kind.py を util.py にリネームし共通 TOML ヘルパーを再利用
`validators/` 配下のファイル名慣例 (config 種類名 = team / workflow / auth /
evaluator / ...) と整合しない `unit_kind.py` を、汎用ヘルパー集約モジュール
として `util.py` にリネーム。あわせて `_detect_unit_kind` 内の `tomllib.load`
直呼びを撤去し、既存の `mixseek.utils.toml.load_toml_with_workspace` を再利用
することで TOML 読み込みパスを一本化。

PR #136 のレビュー指摘に対応:
- unit_kind.py: validators/ 配下のファイル名は config 名を表す慣例なので
  util.py 等の方が良い (#3144661991)
