+++ src/mixseek/config/preflight/validators/__init__.py
@@ -13,7 +13,7 @@ from mixseek.config.preflight.validators.judgment import _validate_judgment
 from mixseek.config.preflight.validators.orchestrator import _validate_orchestrator
 from mixseek.config.preflight.validators.prompt_builder import _validate_prompt_builder
 from mixseek.config.preflight.validators.team import _validate_teams
-from mixseek.config.preflight.validators.unit_kind import _detect_unit_kind
+from mixseek.config.preflight.validators.util import _detect_unit_kind
 from mixseek.config.preflight.validators.workflow import _validate_workflows
 from mixseek.config.preflight.validators.workspace import _validate_workspace_writable
